Scene components for geometry assets, oragnisation of scene components, continuation of conforming engine to new assets api

+++ FoolsEngine/src/FoolsEngine/Renderer/1 - Primitives/VertexData.h
@@ -28,7 +28,7 @@ namespace fe
 
 			ShaderData::Primitive Primitive() const { return ShaderData::PrimitiveInType(Type); }
 			ShaderData::Structure Structure() const { return ShaderData::StructureInType(Type); }
-			uint32_t Size() const { return ShaderData::SizeOfType(Type); }
+			size_t Size() const { return ShaderData::SizeOfType(Type); }
 		};
 
 		class Layout
